Abstract self contained helpers into separate class
